GenPres.Business/Domain/Prescriptions/Prescription.cs
using System;
using GenPres.Business.Data.IRepositories;
using GenPres.Business.Domain.Patients;
using GenPres.Business.Domain.Units;
using GenPres.Business.WebService;

namespace GenPres.Business.Domain.Prescriptions
{
    public class Prescription 
    {   
        public Prescription()
        {
            //Drug = new DrugBo();
        }

        public virtual Patient Patient { get; set; }

        public virtual DateTime StartDate { get; set; }

        public virtual bool IsNew { get { return (Id.ToString() != ""); } }

        public virtual Guid Id
        {
            get; set;
        }

        public static Prescription NewPrescription()
        {
            return new Prescription();
        }
    }
}
